% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effCont.R
\name{effCont}
\alias{effCont}
\title{Continuous Effectiveness Distributions}
\description{
Families to model effectiveness distributions with continuous support. Currently implemented
families are:
\tabular{rl}{
  \code{\link{effCont_norm}} \tab Truncated Normal. \cr
  \code{\link{effCont_beta}} \tab Beta. \cr
  \code{\link{effCont_nks}} \tab Truncated Kernel-smoothed with Gaussian kernel. \cr
  \code{\link{effCont_bks}} \tab Kernel-smoothed with Beta kernel.
}
}
\seealso{
\code{\link{effContFit}} to fit continuous distributions, and
  \code{\link[=eff.cont-class]{eff.cont}} for the S3 class.

  For discrete distributions, see \code{\link{effDisc}}.
}
