% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effCont_beta.R
\name{effCont_beta}
\alias{effCont_beta}
\title{Continuous Effectiveness as Beta Distribution.}
\usage{
effCont_beta(x)
}
\arguments{
\item{x}{a sample of effectiveness scores between 0 and 1.}
}
\value{
an object of class \code{eff.cont.beta}, which inherits from
  \code{\link[=eff.cont-class]{eff.cont}}.
}
\description{
Fits a Beta distribution to the given sample of scores.
}
\examples{
e <- effCont_beta(web2010ap[,1])
c(e$mean, e$var)
plot(e, plot.data = TRUE)
}
\seealso{
\code{\link{deff}}, \code{\link{peff}}, \code{\link{qeff}} and \code{\link{reff}}.
}
