% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effCont_nks.R
\name{effCont_nks}
\alias{effCont_nks}
\title{Continuous Effectiveness as Truncated Gaussian Kernel-smoothed Distribution.}
\usage{
effCont_nks(x)
}
\arguments{
\item{x}{a sample of effectiveness scores between 0 and 1.}
}
\value{
an object of class \code{eff.cont.nks}, which inherits from
  \code{\link[=eff.cont-class]{eff.cont}}.
}
\description{
Fits a kernel-smoothed distribution to the given sample of scores, truncated between 0 and 1, and
using a gaussian kernel.
}
\examples{
e <- effCont_nks(web2010ap[,1])
c(e$mean, e$var)
plot(e, plot.data = TRUE)
}
\seealso{
\code{\link{deff}}, \code{\link{peff}}, \code{\link{qeff}} and \code{\link{reff}}.
}
