% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effCont_norm.R
\name{effCont_norm}
\alias{effCont_norm}
\title{Continuous Effectiveness as Truncated Normal Distribution.}
\usage{
effCont_norm(x)
}
\arguments{
\item{x}{a sample of effectiveness scores between 0 and 1.}
}
\value{
an object of class \code{eff.cont.norm}, which inherits from
  \code{\link[=eff.cont-class]{eff.cont}}.
}
\description{
Fits a Normal distribution, truncated between 0 and 1, to the given sample of scores.
}
\examples{
e <- effCont_norm(web2010ap[,1])
c(e$mean, e$var)
plot(e, plot.data = TRUE)
}
\seealso{
\code{\link{deff}}, \code{\link{peff}}, \code{\link{qeff}} and \code{\link{reff}}.
}
