% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effDisc.R
\name{effDisc}
\alias{effDisc}
\title{Discrete Effectiveness Distributions}
\description{
Families to model effectiveness distributions with discrete support. Currently implemented
families are:
\tabular{rl}{
  \code{\link{effDisc_bbinom}} \tab Beta-Binomial \cr
  \code{\link{effDisc_dks}} \tab Kernel-smoothed with Discrete kernel.
}
}
\seealso{
\code{\link{effDiscFit}} to fit discrete distributions, and
  \code{\link[=eff.disc-class]{eff.disc}} for the S3 class. For continuous distributions, see
  \code{\link{effCont}}.
}
