% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effDisc_bbinom.R
\name{effDisc_bbinom}
\alias{effDisc_bbinom}
\title{Discrete Effectiveness as Beta-Binomial Distribution.}
\usage{
effDisc_bbinom(x, support)
}
\arguments{
\item{x}{a sample of effectiveness scores between 0 and 1.}

\item{support}{the support of the distribution.}
}
\value{
an object of class \code{eff.disc.bbinom}, which inherits from
  \code{\link[=eff.disc-class]{eff.disc}}.
}
\description{
Fits a discrete kernel-smoothed distribution, to the given sample of scores and support points.
}
\examples{
e <- effDisc_bbinom(web2010p20[,1], seq(0,1,.05))
c(e$mean, e$var)
plot(e, plot.data = TRUE)
}
\seealso{
\code{\link{deff}}, \code{\link{peff}}, \code{\link{qeff}} and \code{\link{reff}}.
}
