% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.eff}
\alias{plot.eff}
\alias{dplot}
\alias{pplot}
\alias{qplot}
\title{Plotting tools for effectiveness distributions}
\usage{
\method{plot}{eff}(x, ..., plot.data = TRUE)

dplot(x, ..., plot.data = TRUE)

pplot(x, ..., plot.data = TRUE)

qplot(x, ..., plot.data = TRUE)
}
\arguments{
\item{x}{the effectiveness distribution to plot.}

\item{...}{other arguments to be passed to graphical functions.}

\item{plot.data}{logical: whether to plot the data used to fit the distribution, if any.}
}
\description{
Plot the density, distribution and quantile functions of an effectiveness distribution. Function
\code{plot} plots all three functions in the same graphics device.
}
\seealso{
\code{\link{plot.eff.cont}} and \code{\link{plot.eff.disc}} for more details.
}
