% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{web2010ap}
\alias{web2010ap}
\alias{web2010p20}
\alias{web2010rr}
\title{TREC 2010 Web Ad hoc track.}
\format{A data frame with 88 columns (systems) and 48 rows (queries).}
\usage{
web2010ap

web2010p20

web2010rr
}
\description{
These are the topic-by-system effectiveness matrices for the 88 systems submitted to the TREC
2010 Web Ad hoc track, evaluated over 48 topics. \code{web2010ap} contains Average Precision
scores, \code{web2010p20} contains Precision at 20 scores, and \code{web2010rr} contains
Reciprocal Rank scores.
}
\references{
C.L.A. Clarke, N. Craswell, I. Soboroff, G.V. Cormack (2010). Overview of the TREC
  2010 Web Track. Text REtrieval Conference.
}
\seealso{
\url{https://trec.nist.gov}
}
\keyword{datasets}
